fix the issue with coords in geometry
