+++ src/geometry/geometry.cpp
@@ -25,6 +25,7 @@
 #include "geometry/coordinate_systems.hpp"
 #include "geometry/geometry.hpp"
 #include "geometry/geometry_defaults.hpp"
+#include "geometry/geometry_utils.hpp"
 #include "phoebus_utils/variables.hpp"
 
 using namespace parthenon::package::prelude;
@@ -32,48 +33,6 @@ using parthenon::Coordinates_t;
 using parthenon::ParArray1D;
 
 namespace Geometry {
-
-namespace impl {
-template <typename Data, typename System>
-void SetGeometryDefault(Data *rc, const System &system) {
-  std::vector<std::string> coord_names = {geometric_variables::cell_coords,
-                                          geometric_variables::node_coords};
-  PackIndexMap imap;
-  auto pack = rc->PackVariables(coord_names, imap);
-  PARTHENON_REQUIRE(imap["g.c.coord"].second >= 0, "g.c.coord exists");
-  PARTHENON_REQUIRE(imap["g.n.coord"].second >= 0, "g.n.coord exists");
-  PARTHENON_REQUIRE(imap["g.c.coord"].second - imap["g.c.coord"].first + 1 == 4,
-                    "g.c.coord has correct shape");
-  PARTHENON_REQUIRE(imap["g.n.coord"].second - imap["g.n.coord"].first + 1 == 4,
-                    "g.n.coord has correct shape");
-  int icoord_c = imap["g.c.coord"].first;
-  int icoord_n = imap["g.n.coord"].first;
-
-  auto lamb = KOKKOS_LAMBDA(const int b, const int k, const int j, const int i,
-                            CellLocation loc) {
-    Real C[NDFULL];
-    int icoord = (loc == CellLocation::Cent) ? icoord_c : icoord_n;
-    system.Coords(loc, b, k, j, i, C);
-    SPACETIMELOOP(mu) pack(b, icoord + mu, k, j, i) = C[mu];
-  };
-  IndexRange ib = rc->GetBoundsI(IndexDomain::entire);
-  IndexRange jb = rc->GetBoundsJ(IndexDomain::entire);
-  IndexRange kb = rc->GetBoundsK(IndexDomain::entire);
-  parthenon::par_for(
-      DEFAULT_LOOP_PATTERN, "SetGeometry::Set Cached data, Cent", DevExecSpace(), 0,
-      pack.GetDim(5) - 1, kb.s, kb.e, jb.s, jb.e, ib.s, ib.e,
-      KOKKOS_LAMBDA(const int b, const int k, const int j, const int i) {
-        lamb(b, k, j, i, CellLocation::Cent);
-      });
-  parthenon::par_for(
-      DEFAULT_LOOP_PATTERN, "SetGeometry::Set Cached data, Corn", DevExecSpace(), 0,
-      pack.GetDim(5) - 1, kb.s, kb.e + 1, jb.s, jb.e + 1, ib.s, ib.e + 1,
-      KOKKOS_LAMBDA(const int b, const int k, const int j, const int i) {
-        lamb(b, k, j, i, CellLocation::Corn);
-      });
-  Kokkos::fence(); // do not let users interact with coords unless meshblock data is set
-}
-} // namespace impl
 
 std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
 
@@ -103,9 +62,12 @@ CoordSysMesh GetCoordinateSystem(MeshData<Real> *rc) {
 
 void SetGeometryBlock(MeshBlock *pmb, ParameterInput *pin) {
   MeshBlockData<Real> *rc = pmb->meshblock_data.Get().get();
+  auto *pparent = rc->GetParentPointer().get();
+  StateDescriptor *pkg = pparent->packages.Get("geometry").get();
+  bool do_defaults = pkg->AllParams().Get("do_defaults", true);
   auto system = GetCoordinateSystem(rc);
   SetGeometry<CoordSysMeshBlock>(rc);
-  impl::SetGeometryDefault(rc, system);
+  if (do_defaults) impl::SetGeometryDefault(rc, system);
 }
 
 template <>
